assessment/Header: use AssessmentComponent -component for dividing between PanEU and FRA
